Refactor: Switched the inner array from bool to byte. Operations work with bits to improve efficiency Refactor: Changed the names of files and packages to follow conventions Add: Comments and TODO segments
