+++ structures/bloom_filter/hash.go
