+++ structures/bloom_filter/params.go
